requirements.txt
# SpeedVQA项目依赖

# 核心深度学习框架
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0

# 图像处理
Pillow>=9.0.0
opencv-python>=4.7.0

# 数据处理
numpy>=1.24.0
pandas>=2.0.0

# 配置管理
PyYAML>=6.0

# 测试框架
pytest>=7.0.0
hypothesis>=6.70.0

# 性能优化
tensorrt>=8.6.0  # 可选，用于TensorRT优化
onnx>=1.14.0
onnxruntime>=1.15.0

# 日志和监控
tensorboard>=2.13.0
wandb>=0.15.0  # 可选

# 工具库
tqdm>=4.65.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# 开发工具
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0